+++ DND HP API/Domain/Abstract/Id.cs
@@ -1,8 +1,8 @@
 ﻿namespace DND_HP_API.Domain.Abstract;
 
-public record Id(int Value, bool IsTemporary = false)
+public record Id(long Value, bool IsTemporary = false)
 {
-    public int Value { get; set; } = Value;
+    public long Value { get; set; } = Value;
     public bool IsTemporary { get; set; } = IsTemporary;
 
     public static Id NewTemporaryId(int value=-1)
